// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: string;
  name: string;
  slug: string;
  layout: (
    | {
        heading: string;
        subheading: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        image: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        heading?: string | null;
        form: string | Form;
        id?: string | null;
        blockName?: string | null;
        blockType: 'contact-us-form';
      }
    | {
        heading: string;
        description?: string | null;
        services?:
          | {
              title: string;
              description: string;
              image?: (string | null) | Media;
              id?: string | null;
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'services-overview';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}